Do not hide /lib and /(s)bin symlinks
Avoid including /lib, /bin and /sbin in the nvidia-files.squash. Those
should be symlinks to the actual dirs in /usr, but when included as
normal dirs in the squashfs, overlayfs will replace actual /lib etc
symlinks with that nvidia-only content.

Fix this issue by packaging /usr/lib in place where /lib (etc) was
requested.